Subclass factory for JSON path services
