+++ services/dynamic/dynamic-yaml.service.js
@@ -1,12 +1,10 @@
 'use strict'
 
-const Joi = require('@hapi/joi')
-const jp = require('jsonpath')
-const { renderDynamicBadge, errorMessages } = require('../dynamic-common')
 const { createRoute } = require('./dynamic-helpers')
-const { BaseYamlService, InvalidResponse } = require('..')
+const jsonPath = require('./json-path')
+const { BaseYamlService } = require('..')
 
-module.exports = class DynamicYaml extends BaseYamlService {
+module.exports = class DynamicYaml extends jsonPath(BaseYamlService) {
   static get category() {
     return 'dynamic'
   }
@@ -21,19 +19,11 @@ module.exports = class DynamicYaml extends BaseYamlService {
     }
   }
 
-  async handle(namedParams, { url, query: pathExpression, prefix, suffix }) {
-    const data = await this._requestYaml({
-      schema: Joi.any(),
+  async _getData({ schema, url, errorMessages }) {
+    return this._requestYaml({
+      schema,
       url,
       errorMessages,
     })
-
-    const values = jp.query(data, pathExpression)
-
-    if (!values.length) {
-      throw new InvalidResponse({ prettyMessage: 'no result' })
-    }
-
-    return renderDynamicBadge({ value: values, prefix, suffix })
   }
 }
